backend/pyproject.toml
[tool.poetry]
name = "echoes-backend"
version = "0.1.0"
description = "Echoes LifeMemoir Backend - 知识图谱构建与回忆录生成"
authors = ["Echoes Team"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

# Web 框架
fastapi = "^0.109.0"
uvicorn = { extras = ["standard"], version = "^0.27.0" }
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"

# LangGraph & LangChain
langgraph = "^0.2.0"
langchain = "^0.3.0"
langchain-community = "^0.3.0"

# LLM
ollama = "^0.4.0"
openai = "^1.0.0"

# 数据库
neo4j = "^5.15.0"
asyncpg = "^0.29.0"

# HTTP 客户端
httpx = "^0.27.0"
aiohttp = "^3.9.0"

# 工具
python-dotenv = "^1.0.0"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.1.0"
ruff = "^0.3.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
process-dialogue = "scripts.process_dialogue:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
